requirements.txt
boto3
pandas
requests
psycopg2-binary
sqlalchemy
pyyaml